ALTAPPS-335: Move Platform to core.domain package
