+++ shared/src/androidMain/kotlin/org/hyperskill/app/core/injection/CommonComponentImpl.kt
@@ -8,8 +8,8 @@ import com.russhwolf.settings.AndroidSettings
 import com.russhwolf.settings.Settings
 import kotlinx.serialization.json.Json
 import org.hyperskill.app.BuildConfig
-import org.hyperskill.app.Platform
 import org.hyperskill.app.core.domain.BuildVariant
+import org.hyperskill.app.core.domain.Platform
 import org.hyperskill.app.core.remote.UserAgentInfo
 import org.hyperskill.app.core.view.mapper.ResourceProvider
 import org.hyperskill.app.core.view.mapper.ResourceProviderImpl
