+++ shared/src/commonMain/kotlin/org/hyperskill/app/profile_settings/injection/ProfileSettingsFeatureBuilder.kt
@@ -1,7 +1,7 @@
 package org.hyperskill.app.profile_settings.injection
 
 import kotlinx.coroutines.flow.MutableSharedFlow
-import org.hyperskill.app.Platform
+import org.hyperskill.app.core.domain.Platform
 import org.hyperskill.app.analytic.domain.interactor.AnalyticInteractor
 import org.hyperskill.app.auth.domain.model.UserDeauthorized
 import org.hyperskill.app.core.presentation.ActionDispatcherOptions
