+++ shared/src/iosMain/kotlin/org/hyperskill/app/core/domain/Platform.kt
@@ -1,4 +1,4 @@
-package org.hyperskill.app
+package org.hyperskill.app.core.domain
 
 import platform.UIKit.UIDevice
